move overviewtable to own file
