+++ src/components/dateplanner/overview/Overview.js
@@ -3,10 +3,11 @@ import { useCallback, useEffect, useState } from "react"
 import { getAllAttendences, getEvalByUsergroup, getOwnUsergroups } from '../../../modules/data/DBConnect'
 
 import DateField from "../attendenceInput/DateField"
-import { StyledEvalTable, StyledOverview, StyledOverviewTable } from "./Overview.styled"
+import { StyledEvalTable, StyledOverview } from "./Overview.styled"
 import EvalDiagram from "./EvalDiagram"
 import { Alert, Blank, Check, Deny, PlusOne } from "../attendenceInput/Terminzusage"
 import { IoReload } from "react-icons/io5"
+import { OverviewTable } from "./OverviewTable"
 
 const Overview = ({ theme }) => {
 
@@ -100,49 +101,7 @@ export const Zusage = ({attendence, plusone, theme}) => {
     }
 }
 
-const OverviewTable = ({attendences, theme}) => {
-    return(
-        <StyledOverviewTable>
-            <thead>
-                <tr>
-                    <th>Termin:</th>
-                    {attendences[0].Attendences.map((att) => {
-                        return(<Header key={att.Fullname} Fullname={att.Fullname} />)
-                    })}
-                </tr>
-            </thead>
-            <tbody>
-                {
-                    attendences.map(event => {
-                        return(
-                            <tr key={event.Event_ID}>
-                                <td><DateField dateprops={event}/></td>
-                                {event.Attendences.map(attendence => {
-                                    return(<td key={attendence.Fullname + event.Event_ID}><Zusage attendence={attendence.Attendence} plusone={attendence.PlusOne} theme={theme}/></td>)
-                                })}
-                            </tr>
-                        )
-                    })
-                }
-            </tbody>
-            </StyledOverviewTable>
-    )
-}
 
-const Header = ({ Fullname }) => {
-
-    const [showTooltip, setShowTooltip] = useState(false)
-
-    const initials = Fullname.split(' ')[0].slice(0, 2) + Fullname.split(' ')[1][0]
-
-    const toggleTooltip = () => {
-        setShowTooltip(!showTooltip)
-    }
-
-    return(
-        <th className={showTooltip ? "Tooltip" : "Header"} onClick={toggleTooltip}>{showTooltip ? Fullname : initials}</th>
-    )
-}
 
 const EvalTable = ({evaluation, theme}) => {
     return(
